DB GUI/backend/requirements.txt
flask
flask_socketio
pyserial
pandas
eventlet